Relocated pureText()
